+++ packages/verify-kit/src/validate/presentation.ts
@@ -1,4 +1,4 @@
-import {IVerifiedDataLegacy, IVerifiedDataOnChain, VerifiedData, IVerifiablePresentation, DataVersions} from '@bloomprotocol/att-comms-kit'
+import {IVerifiedDataLegacy, IVerifiedDataOnChain, VerifiedData, IVerifiablePresentation, DataVersions} from '@bloomprotocol/attestations-common'
 
 import {getDecodedTxEventLogs, getDecodedLogValueByName, TDecodedLog} from '../txUtils'
 import {validateVerifiablePresentation} from './structure/presentation'